+++ ui_qt/dialogs/__init__.py
@@ -4,11 +4,13 @@ Dialog windows for PyQt6 UI.
 from ui_qt.dialogs.settings_dialog import SettingsDialog
 from ui_qt.dialogs.hotkey_dialog import HotkeyDialog
 from ui_qt.dialogs.model_manager_dialog import ModelManagerDialog
+from ui_qt.dialogs.model_details_dialog import ModelDetailsDialog
 from ui_qt.dialogs.cleanup_prompt_dialog import CleanupPromptDialog
 
 __all__ = [
     "SettingsDialog",
     "HotkeyDialog",
     "ModelManagerDialog",
+    "ModelDetailsDialog",
     "CleanupPromptDialog",
 ]